requirements.txt
absl-py==0.14.0
asgiref==3.4.1
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
clang==5.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
fastapi==0.68.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h11==0.12.0
h5py==3.1.0
httptools==0.2.0
idna==3.2
imageio==2.9.0
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
networkx==2.6.3
numpy @ file:///C:/ci/numpy_and_numpy_base_1626271900803/work
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.3.3
Pillow==8.3.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.3
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn @ file:///C:/ci/scikit-learn_1622739499047/work
scipy @ file:///C:/ci/scipy_1630606917240/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
starlette==0.14.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tifffile==2021.8.30
typing-extensions==3.7.4.3
urllib3==1.26.7
uuid==1.30
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
